Spectacle Cleanup
- Moves a bunch of stuff out of main() and into other source files.
- Renames remaining SillyShooter files to Specatacle names.
