+++ demos/Spectacle/include/Spectacle.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <Gunship.h>
+
+void MakePlayer( GameObject& player, GameObject& target );
+void MakeCamera( GameObject& camera, GameObject& cameraAnchor, GameObject& player );
+void MakeBullet();
+void MakeEnemy();
