--- demos/Spectacle/include/sillyShooter.h
@@ -1,1 +1,0 @@
-#pragma once
